Unify everything into a single package.
No more core/bench division, we can use the JMH plugin to achieve the same goal, but also make everything simpler.
