+++ src/main/java/com/nickrobison/tdtree/LeafNode.java
