+++ src/main/java/com/nickrobison/tdtree/TDTreeHelpers.java
