--- tdtree-bench/build.gradle
@@ -1,23 +1,0 @@
-plugins {
-    id 'application'
-    id 'com.github.johnrengelman.shadow' version '5.2.0'
-}
-
-ext {
-    jmhVersion = '1.22'
-}
-
-application {
-    mainClassName = "org.openjdk.jmh.Main"
-}
-
-shadowJar {
-    archiveFileName = 'microbenchmarks.jar'
-}
-
-dependencies {
-    compile project(":tdtree")
-    compile("org.openjdk.jmh:jmh-core:${jmhVersion}")
-    compile("org.slf4j:slf4j-simple:1.7.30")
-    annotationProcessor("org.openjdk.jmh:jmh-generator-annprocess:${jmhVersion}")
-}
