--- tdtree/build.gradle
@@ -1,32 +1,0 @@
-plugins {
-    id 'org.sonarqube' version '2.8'
-}
-
-ext {
-    junitVersion = '5.6.1'
-}
-
-apply plugin: 'java-library'
-apply plugin: 'jacoco'
-
-dependencies {
-    implementation group: 'com.nickrobison', name: 'fasttuple-core', version: '0.3.1'
-    implementation group: 'org.slf4j', name: 'slf4j-api', version: '1.7.30'
-    implementation group: 'org.apache.commons', name: 'commons-math3', version: '3.6.1'
-    implementation group: 'it.unimi.dsi', name: 'fastutil', version: '8.3.1'
-    implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.10'
-    compileOnly group: 'org.checkerframework', name: 'checker-qual', version: '3.3.0'
-
-    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: junitVersion
-    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: junitVersion
-}
-
-test {
-    useJUnitPlatform()
-}
-
-jacocoTestReport {
-    reports {
-        xml.enabled = true
-    }
-}
